Initial commit
Purpose:
The purpose of this assignment is to create a custom String class in the form of a singly linked list of characters. This will increase your familiarity with classes, linked lists, and how to overload existing methods. As a result of this assignment, you will learn how to design and create a linked list data structure, and understand in which situations linked lists can be used. The knowledge you gain from this assignment will help you later in this course, other Computer Science courses you might take in the future, and in the Computer Science industry.

Task:

Implement the class SLLString using the SLLString.h file [Download]Download [Download]
Run the program with the help of the Main.cpp file [Download]Download [Download]
To Do:

Download the SLLString.h and Main.cpp files
Use a singly linked list to implement string as a linked list of characters
Apply the Rule of Three by implementing these methods:
destructor
copy constructor
copy assignment operator
Overload the “<<” operator so that std::cout << works with SLLString objects
Implement the support methods
Submit the relevant .cpp and .h files to Canvas